fix: switch test to unix only
